requirements.txt
langchain>=0.1.0
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
pandas>=1.0.0
streamlit>=1.0.0
python-dotenv>=0.15.0
schedule>=1.0.0
shopifyapi>=12.0.0
jupyter>=1.0.0
pytest>=7.0.0